Update kotlin version and revise code syntax
Updated the Kotlin version to 1.9.23 in the project. The build.gradle files and Podfile.lock were updated accordingly. Some syntax changes were made in several files concerning the particles in the Yahtzee view, including refactoring and reducing some lines of code, and removing a deprecated parameter. The score limit was also introduced in the high scores list in the Yahtzee screen.